requirements.txt
# Take pywaggle[audio] to minimize packages to be installed
pywaggle[audio]==0.56.*
